# Q6_K AVX improvements