- [x] I have read the [contributing guidelines](https://github.com/ggerganov/llama.cpp/blob/master/CONTRIBUTING.md)
- Self-reported review complexity:
  - [ ] Low
  - [x] Medium
  - [ ] High

I've updated the Q6_K AVX implementation for better prompt processing and text generation speed. Most of the improvements came from using the `bsums` data to separately calculate the Q6 offsets (it's precomputed for a reason!), reducing the number of instructions.

The same techniques should also be applicable to the AVX2 implementation if someone wants to work on that. It looks like the ARM implementation already uses `bsums` correctly.

By the way I'm kinda curious if anyone here actually uses llama.cpp on an AVX only machine, or if I'm the only one :disappointed:.

**Results (Xeon E5 v2):**

Master
```
q6_K
  vec_dot_q
    4096 values (0.02 MB)
      min cycles/32 vals   :     20.29
      avg cycles/32 vals   :     20.36
      float32 throughput   :     15.26 GB/s
      quantized throughput :      3.13 GB/s
```
PR
```
q6_K
  vec_dot_q
    4096 values (0.02 MB)
      min cycles/32 vals   :     16.43
      avg cycles/32 vals   :     16.51
      float32 throughput   :     19.07 GB/s
      quantized throughput :      3.91 GB/s
```

| model                          |       size |     params | backend    | threads |          test |                  t/s | speedup |
| ------------------------------ | ---------: | ---------: | ---------- | ------: | ------------: | -------------------: |-------------------: |
| llama 7B Q6_K (Master)                  |   5.53 GiB |     7.24 B | CPU        |       8 |         pp512 |          7.59 ± 0.02 | |
| llama 7B Q6_K (Master)                 |   5.53 GiB |     7.24 B | CPU        |       8 |         tg128 |          5.75 ± 0.01 | |
| llama 7B Q6_K (PR)                  |   5.53 GiB |     7.24 B | CPU        |       8 |         pp512 |         10.29 ± 0.03 | 36% |
| llama 7B Q6_K (PR)                  |   5.53 GiB |     7.24 B | CPU        |       8 |         tg128 |          6.50 ± 0.02 | 13% |

`test-backend-ops` and `test-quantize-fns` is passing.