@slaren 2024-11-01
I get these results with `test-backend-ops perf`:

master:
```
  MUL_MAT(type_a=q6_K,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                  32706 runs -   305.94 us/run - 117.44 MFLOP/run - 383.87 GFLOPS
  MUL_MAT(type_a=q6_K,type_b=f32,m=4096,n=512,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                   87 runs - 115503.36 us/run -  60.13 GFLOP/run - 520.59 GFLOPS
```

PR:
```
  MUL_MAT(type_a=q6_K,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                  37812 runs -   264.88 us/run - 117.44 MFLOP/run - 443.38 GFLOPS
  MUL_MAT(type_a=q6_K,type_b=f32,m=4096,n=512,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                  107 runs - 94010.70 us/run -  60.13 GFLOP/run - 639.60 GFLOPS
```