@Nexesenex 2024-11-01
> The same techniques should also be applicable to the AVX2 implementation if someone wants to work on that. It looks like the ARM implementation already uses bsums correctly.

Considering that most >3bpw quants use an output weight quantized in Q6_K (the Q6_K FTYPE using a Q6_K token embed weight as well), and that not offloading on GPU the output weight is done sometimes to save a bit of VRAM, would such AVX2 implementation reduce the loss of performances incurred when not offloading said output weight on GPU?

I'm using a Ryzen 5700G, which is AVX2 I believe.