@netrunnereve 2024-11-01
> Considering that most >3bpw quants use an output weight quantized in Q6_K (the Q6_K FTYPE using a Q6_K token embed weight as well), and that not offloading on GPU the output weight is done sometimes to save a bit of VRAM, would such AVX2 implementation reduce the loss of performances incurred when not offloading said output weight on GPU?

I mean it should help in theory, but considering how the output weight is only a small part of the full model you probably won't notice a thing. I guess the easiest way to test this is to benchmark a Q4 or Q5 model with this PR.