@netrunnereve 2024-11-04
I had to update the CI target macOS-latest-cmake-x64 to run on macos-13 as macos-12 seems to have been depreciated by Github. The macos-13 machines are still x86 so this won't generate arm binaries (see #7940).