@SmallAndSoft 2024-11-05
> By the way I'm kinda curious if anyone here actually uses llama.cpp on an AVX only machine, or if I'm the only one 😞.

I do. AMD FX is AVX only.